+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes qsf-integration project will be documented in this file.
 ### Deprecated
 
 ### Fixed
+- fix test execution by updating maven-surefire-plugin
 
 ## [2.16]
 ### Added