+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes qsf-integration project will be documented in this file.
 - implement multi select faceting for sliders
 - add a data structure for redirects
 - implement a DocumentStringSubstitutor to replace all variables in a string with the values of the document
+- make it possible to override the loading of a profile
 
 ### Changed
 - use StringSubstitutor instead of StrSubstitutor because it is deprecated